% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{getdepthmap}
\alias{getdepthmap}
\title{read in image and change them into a depth map}
\usage{
getdepthmap(path, cutoff = 0.5)
}
\arguments{
\item{path}{path of the image}

\item{cutoff}{cutoff the grayscale, from 0 to 1}
}
\description{
read in image and change them into a depth map
}
