% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getmagiceye}
\alias{getmagiceye}
\title{Generate autosterograms}
\usage{
getmagiceye(x, y, shift)
}
\arguments{
\item{x}{depth matrix}

\item{y}{pattern matrix}

\item{shift}{shift for magic eye}
}
\value{
matrix of autosterograms
}
\description{
Generate autosterograms
}
