% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{getme}
\alias{getme}
\title{read in image and change them into magic eye}
\usage{
getme(img, shift = 0.15)
}
\arguments{
\item{img}{cimg object}

\item{shift}{shift of magic eye, default 0.15}
}
\description{
read in image and change them into magic eye
}
