% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magiceyer.R
\docType{package}
\name{magiceyer}
\alias{magiceyer}
\title{magiceyer: Creat Autostereograms in R}
\description{
Using your own picture to create autosterograms in R. You could aslo generate magic eye with text.
}
